move ColumnVector.Array and ColumnarBatch.Row to individual files
