+++ sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/ColumnVector.java
@@ -16,11 +16,9 @@
  */
 package org.apache.spark.sql.execution.vectorized;
 
-import org.apache.spark.sql.catalyst.InternalRow;
-import org.apache.spark.sql.catalyst.util.ArrayData;
 import org.apache.spark.sql.catalyst.util.MapData;
-import org.apache.spark.sql.types.*;
-import org.apache.spark.unsafe.types.CalendarInterval;
+import org.apache.spark.sql.types.DataType;
+import org.apache.spark.sql.types.Decimal;
 import org.apache.spark.unsafe.types.UTF8String;
 
 /**
@@ -42,190 +40,6 @@ import org.apache.spark.unsafe.types.UTF8String;
  * ColumnVectors are intended to be reused.
  */
 public abstract class ColumnVector implements AutoCloseable {
-
-  /**
-   * Holder object to return an array. This object is intended to be reused. Callers should
-   * copy the data out if it needs to be stored.
-   */
-  public static final class Array extends ArrayData {
-    // The data for this array. This array contains elements from
-    // data[offset] to data[offset + length).
-    public final ColumnVector data;
-    public int length;
-    public int offset;
-
-    // Populate if binary data is required for the Array. This is stored here as an optimization
-    // for string data.
-    public byte[] byteArray;
-    public int byteArrayOffset;
-
-    // Reused staging buffer, used for loading from offheap.
-    protected byte[] tmpByteArray = new byte[1];
-
-    protected Array(ColumnVector data) {
-      this.data = data;
-    }
-
-    @Override
-    public int numElements() { return length; }
-
-    @Override
-    public ArrayData copy() {
-      throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public boolean[] toBooleanArray() { return data.getBooleans(offset, length); }
-
-    @Override
-    public byte[] toByteArray() { return data.getBytes(offset, length); }
-
-    @Override
-    public short[] toShortArray() { return data.getShorts(offset, length); }
-
-    @Override
-    public int[] toIntArray() { return data.getInts(offset, length); }
-
-    @Override
-    public long[] toLongArray() { return data.getLongs(offset, length); }
-
-    @Override
-    public float[] toFloatArray() { return data.getFloats(offset, length); }
-
-    @Override
-    public double[] toDoubleArray() { return data.getDoubles(offset, length); }
-
-    // TODO: this is extremely expensive.
-    @Override
-    public Object[] array() {
-      DataType dt = data.dataType();
-      Object[] list = new Object[length];
-      try {
-        for (int i = 0; i < length; i++) {
-          if (!data.isNullAt(offset + i)) {
-            list[i] = get(i, dt);
-          }
-        }
-        return list;
-      } catch(Exception e) {
-        throw new RuntimeException("Could not get the array", e);
-      }
-    }
-
-    @Override
-    public boolean isNullAt(int ordinal) { return data.isNullAt(offset + ordinal); }
-
-    @Override
-    public boolean getBoolean(int ordinal) {
-      return data.getBoolean(offset + ordinal);
-    }
-
-    @Override
-    public byte getByte(int ordinal) { return data.getByte(offset + ordinal); }
-
-    @Override
-    public short getShort(int ordinal) {
-      return data.getShort(offset + ordinal);
-    }
-
-    @Override
-    public int getInt(int ordinal) { return data.getInt(offset + ordinal); }
-
-    @Override
-    public long getLong(int ordinal) { return data.getLong(offset + ordinal); }
-
-    @Override
-    public float getFloat(int ordinal) {
-      return data.getFloat(offset + ordinal);
-    }
-
-    @Override
-    public double getDouble(int ordinal) { return data.getDouble(offset + ordinal); }
-
-    @Override
-    public Decimal getDecimal(int ordinal, int precision, int scale) {
-      return data.getDecimal(offset + ordinal, precision, scale);
-    }
-
-    @Override
-    public UTF8String getUTF8String(int ordinal) {
-      return data.getUTF8String(offset + ordinal);
-    }
-
-    @Override
-    public byte[] getBinary(int ordinal) {
-      return data.getBinary(offset + ordinal);
-    }
-
-    @Override
-    public CalendarInterval getInterval(int ordinal) {
-      int month = data.getChildColumn(0).getInt(offset + ordinal);
-      long microseconds = data.getChildColumn(1).getLong(offset + ordinal);
-      return new CalendarInterval(month, microseconds);
-    }
-
-    @Override
-    public InternalRow getStruct(int ordinal, int numFields) {
-      return data.getStruct(offset + ordinal);
-    }
-
-    @Override
-    public ArrayData getArray(int ordinal) {
-      return data.getArray(offset + ordinal);
-    }
-
-    @Override
-    public MapData getMap(int ordinal) {
-      throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Object get(int ordinal, DataType dataType) {
-      if (dataType instanceof BooleanType) {
-        return getBoolean(ordinal);
-      } else if (dataType instanceof ByteType) {
-        return getByte(ordinal);
-      } else if (dataType instanceof ShortType) {
-        return getShort(ordinal);
-      } else if (dataType instanceof IntegerType) {
-        return getInt(ordinal);
-      } else if (dataType instanceof LongType) {
-        return getLong(ordinal);
-      } else if (dataType instanceof FloatType) {
-        return getFloat(ordinal);
-      } else if (dataType instanceof DoubleType) {
-        return getDouble(ordinal);
-      } else if (dataType instanceof StringType) {
-        return getUTF8String(ordinal);
-      } else if (dataType instanceof BinaryType) {
-        return getBinary(ordinal);
-      } else if (dataType instanceof DecimalType) {
-        DecimalType t = (DecimalType) dataType;
-        return getDecimal(ordinal, t.precision(), t.scale());
-      } else if (dataType instanceof DateType) {
-        return getInt(ordinal);
-      } else if (dataType instanceof TimestampType) {
-        return getLong(ordinal);
-      } else if (dataType instanceof ArrayType) {
-        return getArray(ordinal);
-      } else if (dataType instanceof StructType) {
-        return getStruct(ordinal, ((StructType)dataType).fields().length);
-      } else if (dataType instanceof MapType) {
-        return getMap(ordinal);
-      } else if (dataType instanceof CalendarIntervalType) {
-        return getInterval(ordinal);
-      } else {
-        throw new UnsupportedOperationException("Datatype not supported " + dataType);
-      }
-    }
-
-    @Override
-    public void update(int ordinal, Object value) { throw new UnsupportedOperationException(); }
-
-    @Override
-    public void setNullAt(int ordinal) { throw new UnsupportedOperationException(); }
-  }
-
   /**
    * Returns the data type of this column.
    */
@@ -350,7 +164,7 @@ public abstract class ColumnVector implements AutoCloseable {
   /**
    * Returns a utility object to get structs.
    */
-  public ColumnarBatch.Row getStruct(int rowId) {
+  public VectorBasedRow getStruct(int rowId) {
     resultStruct.rowId = rowId;
     return resultStruct;
   }
@@ -359,7 +173,7 @@ public abstract class ColumnVector implements AutoCloseable {
    * Returns a utility object to get structs.
    * provided to keep API compatibility with InternalRow for code generation
    */
-  public ColumnarBatch.Row getStruct(int rowId, int size) {
+  public VectorBasedRow getStruct(int rowId, int size) {
     resultStruct.rowId = rowId;
     return resultStruct;
   }
@@ -367,7 +181,7 @@ public abstract class ColumnVector implements AutoCloseable {
   /**
    * Returns the array at rowid.
    */
-  public final ColumnVector.Array getArray(int rowId) {
+  public final VectorBasedArray getArray(int rowId) {
     resultArray.length = getArrayLength(rowId);
     resultArray.offset = getArrayOffset(rowId);
     return resultArray;
@@ -376,7 +190,7 @@ public abstract class ColumnVector implements AutoCloseable {
   /**
    * Loads the data into array.byteArray.
    */
-  public abstract void loadBytes(ColumnVector.Array array);
+  public abstract void loadBytes(VectorBasedArray array);
 
   /**
    * Returns the value for rowId.
@@ -423,12 +237,12 @@ public abstract class ColumnVector implements AutoCloseable {
   /**
    * Reusable Array holder for getArray().
    */
-  protected ColumnVector.Array resultArray;
+  protected VectorBasedArray resultArray;
 
   /**
    * Reusable Struct holder for getStruct().
    */
-  protected ColumnarBatch.Row resultStruct;
+  protected VectorBasedRow resultStruct;
 
   /**
    * The Dictionary for this column.
