+++ sql/core/src/main/java/org/apache/spark/sql/execution/vectorized/OnHeapColumnVector.java
@@ -494,7 +494,7 @@ public final class OnHeapColumnVector extends WritableColumnVector {
   }
 
   @Override
-  public void loadBytes(ColumnVector.Array array) {
+  public void loadBytes(VectorBasedArray array) {
     array.byteArray = byteData;
     array.byteArrayOffset = array.offset;
   }
